java 8 version of interface implemented

+++ src/main/java/com/kawnayeen/java8/_why_java_8/IPersonRepository.java
@@ -24,5 +24,5 @@ public interface IPersonRepository {
 
     Person findOldestFemale(List<Person> personList);
 
-    int averageAge(List<Person> personList);
+    double averageAge(List<Person> personList);
 }
