+++ src/main/java/com/kawnayeen/java8/_why_java_8/_java_7/Java7PersonRepository.java
@@ -53,7 +53,7 @@ public class Java7PersonRepository implements IPersonRepository {
     }
 
     @Override
-    public int averageAge(List<Person> personList) {
+    public double averageAge(List<Person> personList) {
         return 0;
     }
 }
